FC_GUI/requirements.txt
accelerate==0.30.1
anyio==4.6.0
async-property==0.2.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
-e git+ssh://git@github.com/mcrts/deepmodelz.git@19ddff93766b5f861198dbeb5e404bc6e1704c7e#egg=deepmodelz
deprecation==2.1.0
docker-pycreds==0.4.0
exceptiongroup==1.2.2
filelock==3.14.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
python-dateutil==2.9.0.post0
python-keycloak==4.5.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.4.3
sentry-sdk==2.2.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.12.1
tokenizers==0.15.2
torch==2.3.1
tqdm==4.66.4
transformers==4.37.2
triton==2.3.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
